+++ python/boomer/seco/stopping_criteria.pxd
@@ -1,4 +1,4 @@
-from boomer.common._arrays cimport intp, float64
+from boomer.common._arrays cimport uint32, float64
 from boomer.common.statistics cimport AbstractStatistics
 from boomer.common.stopping_criteria cimport StoppingCriterion
 
@@ -11,4 +11,4 @@ cdef class UncoveredLabelsCriterion(StoppingCriterion):
 
     # Functions:
 
-    cdef bint should_continue(self, AbstractStatistics* statistics, intp num_rules)
+    cdef bint should_continue(self, AbstractStatistics* statistics, uint32 num_rules)
